Move userlist and userprofile scene
