+++ src/features/user/screens/UserListScene.js
@@ -1,25 +1,18 @@
 // @flow
 import React, {Component} from 'react';
-import {View} from 'react-native';
-import {Text} from '../../core-ui';
-import {USER_LIST} from '../../data/users';
-import UserCard from '../../components/UserCard';
+import {View, StyleSheet} from 'react-native';
+import {Text} from '../../../core-ui';
+import {USER_LIST} from '../../../data/users';
+import UserCard from '../../../components/UserCard';
 
 export default class UserList extends Component<{}> {
   render() {
     return (
-      <View
-        style={{
-          flex: 1,
-          paddingVertical: 20,
-        }}
-      >
-        <Text size="xlarge" style={{paddingHorizontal: 10}}>
+      <View style={styles.root}>
+        <Text size="xlarge" style={styles.title}>
           User List
         </Text>
-        <View
-          style={{flexDirection: 'row', flexWrap: 'wrap', alignItems: 'center'}}
-        >
+        <View style={styles.content}>
           {USER_LIST.map((user, index) => {
             let {name, username, website, phone} = user;
             return (
@@ -37,3 +30,18 @@ export default class UserList extends Component<{}> {
     );
   }
 }
+
+const styles = StyleSheet.create({
+  root: {
+    flex: 1,
+    paddingVertical: 20,
+  },
+  title: {
+    paddingHorizontal: 10,
+  },
+  content: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    alignItems: 'center',
+  },
+});
