+++ src/features/user/screens/UserProfileScene.js
@@ -1,12 +1,12 @@
 // @flow
 import React, {Component} from 'react';
 import {View} from 'react-native';
-import {Text} from '../../core-ui';
+import {Text} from '../../../core-ui';
 
 export default class UserDetailScene extends Component<{}> {
   render() {
     return (
-      <View>
+      <View style={{paddingHorizontal: 12, paddingVertical: 5}}>
         <Text>Hello this is text</Text>
       </View>
     );
